deprecated course-chart-detail and migrated over to the course-chart component
