--- src/app/course-chart-detail/course-chart-detail.component.html
@@ -1,6 +1,0 @@
-<div>
-  <div class="course-chart">
-    <div id="chart_div_0" class="charts"></div>
-    <div id="chart_div_1" class="charts"></div>
-  </div>
-</div>
