+++ src/app/course-chart/course-chart.component.css
@@ -16,14 +16,11 @@
   width: 100%;
 }
 .course-chart {
-  padding: 20px;
   width: 100%;
 }
 .course-chart .charts {
-  width: 50%;
-  display: inline-block;
-  vertical-align: top;
-  height: 300px;
+  height: 500px;
+  margin-bottom: 20px;
 }
 .course-chart rect {
   fill: transparent;
