+++ src/app/course-chart/course-chart.component.html
@@ -1,8 +1,11 @@
-<div *ngIf="data" class="course-chart-component">
+<div *ngIf="chart_data" class="course-chart-component">
   <h2 class="title">COURSE DATA</h2>
   <div class="scroll-zone-container">
     <div class="scroll-zone">
-      <app-course-chart-detail [data]="data"></app-course-chart-detail>
+      <div class="course-chart">
+        <div id="chart_div_0" class="charts"></div>
+        <div id="chart_div_1" class="charts"></div>
+      </div>
     </div>
   </div>
 </div>
