+++ src/app/course-chart/course-chart.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input }   from '@angular/core';
+import { Component, OnInit }   from '@angular/core';
 import { Resort }                     from '../resort';
 import { PlotlyData }                 from '../plotly-data';
 import { ResortService }              from '../resort.service';
@@ -10,18 +10,36 @@ import { ResortService }              from '../resort.service';
 })
 export class CourseChartComponent implements OnInit {
   resorts: Resort[];
-  data: Array<any>;
+  chart_data: Array<any>;
+  charts: any[];
+  isChartInitialized: boolean;
 
   constructor(private resortService: ResortService) {
-    this.data = [];
-    /*
-    this issue is that I need an event or set of logic that fires after the section has been triggered.  the google charts need to initialized and loaded AFTER we switch to this panel/section.  There's no known event for that.  ngOnInit() may not be good becasue I'm not using an router....  (maybe)
-    */
+    this.resorts = [];
+    this.chart_data = [];
+    this.charts = [];
+    this.isChartInitialized = false;
   }
 
   ngOnInit() {
     this.resorts = this.resortService.getResorts();
     this.setDataFromResorts();
+  }
+
+  ngOnChanges(): void {
+    this.resorts = this.resortService.getResorts();
+    this.setDataFromResorts();
+    // semi-exhaustive check to avoid premature calls
+    if(this.isChartInitialized && this.charts && this.charts.length > 0 && this.chart_data && this.chart_data.length > 0) {
+      this.updateCharts();
+    }
+  }
+
+  ngAfterViewChecked(): void {
+    // I really hate this mechanism
+    if(!this.isChartInitialized) {
+      this.loadCharts();
+    }
   }
 
   setDataFromResorts(): void {
@@ -46,12 +64,31 @@ export class CourseChartComponent implements OnInit {
     let data_0 = [data_trails, data_lifts];
     let data_1 = [data_acres];
 
-    this.data = [data_0, data_1];
+    this.chart_data = [data_0, data_1];
   }
 
-  // may need throttling
-  ngOnChanges(): void {
-    this.resorts = this.resortService.getResorts();
-    this.setDataFromResorts();
+  loadCharts(): void {
+    // need to read from data, init charts, and store chart instances to this.charts
+    if(this.chart_data && this.chart_data.length > 0) {
+      let data_0 = this.chart_data[0];
+      let data_1 = this.chart_data[1];
+      let layout_0 = {title: 'Number of Trails and Lifts per Resort', barmode: 'group'};
+      let layout_1 = {title: 'Number of Acres per Resort', barmode: 'group'};
+
+      let chart_0 = Plotly.newPlot('chart_div_0', data_0, layout_0);
+      let chart_1 = Plotly.newPlot('chart_div_1', data_1, layout_1);
+
+      this.isChartInitialized = true;
+      this.charts = [chart_0,chart_1];
+    } else {
+      console.log("COURSE-CHART-DETAIL: NO DATA AND NO CHARTS?", this.chart_data, this.charts);
+    }
+  }
+
+  updateCharts(): void {
+    // eventually, this update should check each chart and only update the with the updated data set
+    if(this.isChartInitialized){
+      this.loadCharts();
+    }
   }
 }
